JDK 25 compatibility - fix jol-core API changes and checkstyle

+++ stats/src/main/java/com/facebook/airlift/stats/cardinality/DenseHll.java
@@ -46,7 +46,7 @@ final class DenseHll
     private static final int MAX_DELTA = (1 << BITS_PER_BUCKET) - 1;
     private static final int BUCKET_MASK = (1 << BITS_PER_BUCKET) - 1;
 
-    private static final int DENSE_INSTANCE_SIZE = ClassLayout.parseClass(DenseHll.class).instanceSize();
+    private static final long DENSE_INSTANCE_SIZE = ClassLayout.parseClass(DenseHll.class).instanceSize();
     private static final int OVERFLOW_GROW_INCREMENT = 5;
 
     private final byte indexBitLength;
